remove cast

+++ src/main/java/leejoongseok/wms/item/domain/ItemSize.java
@@ -36,6 +36,6 @@ public class ItemSize {
     }
 
     public Long calculateVolume() {
-        return (long) ((long) widthMillimeter * lengthMillimeter * heightMillimeter);
+        return (long) widthMillimeter * lengthMillimeter * heightMillimeter;
     }
 }
